backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
pandas==1.3.3
python-dotenv==0.19.0
werkzeug==2.0.1
sqlalchemy==1.4.23
openpyxl==3.0.9
requests==2.26.0
xmltodict==0.13.0
numpy==1.24.3
scikit-learn==1.3.0 